+++ dflib-excel/src/test/java/com/nhl/dflib/excel/BaseExcelTest.java
@@ -1,0 +1,15 @@
+package com.nhl.dflib.excel;
+
+import org.junit.jupiter.api.io.TempDir;
+
+import java.io.File;
+
+public class BaseExcelTest {
+
+    @TempDir
+    static File OUT_BASE;
+
+    protected static String outPath(String name) {
+        return OUT_BASE.getPath() + File.separator + name;
+    }
+}